+++ NEWS.md
@@ -1,3 +1,8 @@
+# isocubes 0.1.3  2022-07-10
+
+* Added a `light` argument to specify lighting direction
+    * renamed `fill_left` and `fill_right` to `fill2` and `fill3` respectively
+
 # isocubes 0.1.2  2022-06-23
 
 * Replaced `max_y` argument with its inverse `ysize`